Implementation of barrel import transformer
test script will first run the barrel import transform step which will find all exports as specified in the

Jest is configured to use the transformer if the transform cache exists.

The custom jest transformer looks at a typescript file and processes it to convert the file by transforming the import and then pass the transformed text to ts-jest to process as can be seen in the barrelImportTransformerProcessor.

More info added to README file.